+++ Source/YTE/Graphics/Generics/Mesh.hpp
@@ -282,8 +282,8 @@ namespace YTE
     Mesh(const std::string &aFile,
          std::vector<Submesh> &aSubmeshes);
 
-    virtual void UpdateVertices(int aSubmeshIndex, std::vector<Vertex>& aVertices);
-    virtual void UpdateVerticesAndIndices(int aSubmeshIndex, std::vector<Vertex>& aVertices, std::vector<u32>& aIndices);
+    virtual void UpdateVertices(size_t aSubmeshIndex, std::vector<Vertex>& aVertices);
+    virtual void UpdateVerticesAndIndices(size_t aSubmeshIndex, std::vector<Vertex>& aVertices, std::vector<u32>& aIndices);
 
     virtual ~Mesh();
